+++ client/src/lib/parse/data/SetParser.ts
@@ -109,7 +109,7 @@ export default class SetParser extends EntityParser {
     null | string | Array<any> | IContentWithLanguage
   > | null {
     const names = this.getNames()
-    const entityClass = this.getEntityClass()
+    const entityClass = this.getEntityClass('work')
     const itemType = this.getTypes()
     const identifiers = this.getIdentifiers()
     const setCreationEvent = this.getCreationEvent()